Center quote and push quote source right
Text spread across entire screen.
Add flex to the contaier named .quote and center through justify-content.
Padding to quote text and source pushes the grey area to the right places.
Flex-wrap to the container puts the quote source on a separate line. Finally, margin-right: auto pushes the quote source to the right place on the bottom.